Using ListViewCard for company-all and implemented search in that page
In Company Tab, a Floating Button was added to present all companies.
The companies are now using the component ListViewCard instead of the
custom previous one. A search icon in the appBar of this new screen was
added to allow the user to search for a company. The search is working
but in terms of design it may change.

Changelog:
 - Add Floating Button in company-table screen
 - Changed CompanyCard Widget to use the Widget from ListViewCard
 - Add search icon in AppBar of company-all screen
 - Implemented search by company in this new screen