Added Sorting Methods in company-all and Add Company Button
Some otimizations were done to improve the perfomance of the company-all
view. First of all, the GridView.count was replaced by a
GridView.builder. Then, a method to companyService was added to get a
List of CompanyLight instead of Company (only includes name, id, image
and participations of companies). Also added a Floating Button to Add a
Company (it will redirect to the Add Company Page later on). Implemented
a button in appBar that display a dropdown with sorting methods (to
sort the companies displayed). These
sorting methods are: sort by number of participations and sort by last
participation.

Changelog:
 - Otimize perfomance of screen company-all
 - Add Company Button added
 - Added sort mechanism to company-all