Reformulated Search, add participation info to card, using ListViewCard
In Company Table, changed Floating Button to be more clear to the user,
by adding a string to the icon. In Company-all screen, added number of
participations and last participation to the card to make the sort
method more clear. The cards now are using ListViewCard component. In
search, the body is now the GridView and while the user puts its query
in the search, the GridView reduces its size to only show the matched
companies.

Changelog:
 - In Company Table, Floating Button now has the string 'Show All
Companies'
 - Using ListViewCard for the cards of company-all
 - Displaying in each card of company-al the number of participations
and last participation
 - Search now uses GridView, like default view