Speaker-all screen and components Grid Layout and Search Delegate
Created the speaker-all screen, which is very similar to the company-all
screen. Fixed a Bug in ListViewCard (so that the speaker-all screen
showed all speakers from previous editions). Created the component Grid
Layout that is shared by company all and speaker all and might be useful
for the members grid layout. The search delegate component might also be
useful for the search in other parts of the app.

Changelog:
 - Speaker-all screen (similar to company-all screen)
 - Component Grid Layout to show company/speakers in grid
 - Component Search used to search for a speaker or company