Fixed duplicated FloatingButton. Change to CompanyLight
In order to improve perfomance, the cards are now using CompanyLight
instead of Company. The CompanyLight includes also the last
participation and participation status. The participation status is
based on the principle that the participations list is ordered according
to the event (fragile principle). Also fixed the bug of having two
Floating Buttons in company-table and speaker-table.

Changelog:
 - Company-Table and Company-all using CompanyLight
 - Took duplicated FloatingButton from company-table and speaker-table