CompanyListWidget now fetches companies gradually
The pagination method implemented on backend to fetch all companies was
improved. We continue to send a limit that sets the maximum number of
companies that we are going to fetch but to fetch more companies, we
send to backend the ID of the last company fetched and we perform a
filter that filters the companies with ID greater than the one we
sent (this is possible because ObjectID have natural order). The
CompanyListWidget now fetches 30 items each time and with
ScrollController when we reach the bottom, it fetches more 30 companies.
The search is something to see yet because we are doing a request to
backend to get all companies and this causes a glitch.

Changelog:
 - Backend: To fetch more companies, we send ID of last company fetched
and we filter the companies that have ID greater than the one we sent
 - Frontend: CompanyListWidget now fetches 30 companies at once and only
loads more companies when we reach the bottom.
 - PROBLEM: Search is glitchy (because we are doing a request to backend
to fetch all companies).