Sorting Methods now using backend to sort
Implemented a new parameter in GET /companies endpoint to tell if
sorting is necessary. This parameter is not null only when there's
sorting (by last participation or number of participations). Used an
aggregation to perform the sorting (to sort by number of participations,
I add a field to each document telling the size of array containing the
multiple participations and then I order the documents by that
parameter). Finally, I skip the documents that I already retrieved to
frontend and set the max number of items coming to frontend.
(Previously, I used the natural order of ID of documents to filter the
one's I wanted but that does not work with sorting methods).

Changelog:
 - Added SortingMethod parameter to endpoint and changed name of
previousID parameter to numRequests (since I no longer pass an ID, but a
number telling the number of requests I did to backend)
 - Implemented aggregation to sort the documents in backend